damage text
added damage text that flies off of the player when hit. This happens at
the beginning of the show when they are first attacked by the hippies.
In the show the number flies upward, slows, and fades. I tried it that
way first, but it didn't look quite right because of the face that in
our version the player rebounds when struck. So I made it so that they
fly off and drop towards the ground, as if affected by gravity.

The damage number is set in the hippy collide function. For furture
enemies, we can do the same.