Made forest2 inventory-friendly, added crafting to inventory.
You can now craft!
Go to the items in the inventory page, and press shift
on one to add it to the crafting table.
Select an item on the crafting table and press shit
to remove it.
When you have two items on the crafting table that can
be crafted together, select the space to the left of the
items on the crafting table and press shift to craft them..
Crafting recipies can be edited from the items/recipes lua
file.
The only recipe that currently exists is stick + rock = throwing knives
This is also the only crafting recipe shown in the show, so for the
rest we're going to have to get creative.