Closes #424 Changed music to "Abed's Fairwell"
As per conversation in reddit, specifically with Operia2,
there actually was some more appropriate music for this
area. So I added that .ogg and set the map property.

I'm leaving the old one in there, since the Forest 2 map
might appreciate the love, but didn't take the liberty of
changing the map to use it. Not my call to make.

http://www.reddit.com/r/hawkthorne/comments/117sc2/suggestion_different_music_for_village_forest/c6ls739