Continue screen, Lives, and Money
I created a continue screen as part of pause.lua that uses the "Quit to
Menu" option as a continue and resets lives/money to 3/0. "Quit to
Desktop" is also an option.

I also added global variables "lives" and "money" to decrement lives on
death and keep money persistant until you run out of lives.

I'm new to Lua/Love and github, so any feedback will be greatly
appreciated. Thanks!