Removes the floor node in favor of wall node
From a .tmx perspective, there is now no difference between objects on the "floor" layer as the "wall" layer
I have retained the ability to use the "floor" layer for backwards compatibility.

Maybe someday when the project is abandoned i'll go back and clean up the layers, but it should have very little effect on performance.