V1
Completed base. Currently the exp required to achieve a level is
calculated by 5(L^2 + L). Enemies now have a variable called exp, which
is how much exp is dropped. The player's exp is stored in variable
called exp. The levels are on the HUD and the progress is above the
health bar.