Made handling of quit option similar to the other actions
Moved the special handling of the quit option from the keypressed
function to its own function, the existing action-handling code will
handle the new quit function by design.

Also added a check for the up/down keypressed events to ensure that the
right-hand menu is selected before parsing the events.