Remove Gamestate in favor of Scenes
All scenes should have test coverage. The new scenes are easy to test,
as evidenved by the spec/scenes_spec.lua file. Since love can't be
included in the tests, I've started work on fakelove, a module that
behaves like love for the tests.

I'm going to contact the Love devs and see if there is a way to load
love using tests.