Don't always post the release thread
After release, we sometimes need to create a bugfix release. Instead of
having to manually do this, make the script that posts the thread and
tweet smart enough to know when we create a bugfix release. To make a
release as bug fux, just include the words "NOSOCIAL" in the body of the
pull request