Fixes picking up items on interactive nodes
This commit will check for non-interactive nodes when using the INTERACT
key. This makes it so that you can pick up items if they are colliding
with an interactive node.

i.e. say you drop an item on an interactive door, if you try to pick up
the item, it may find the door first, and pass the keypressed event to
it, rather than the item. Making it impossible to pick up an item on the
door.