Changes club to switch
On request of the fantastic fellas at the Dunnshire, but I felt it
needed a change, too (When I hear "club", I think of the caveman type)

Puttin' my "Where's Waldo" skills to use to find all instances of
"club". Would have been easier if the words were red and white, though.