Save state of White Crystal
After the user has placed the White Crystal in Hawkthorne Entrance, keep
the door open permanently. This make it more convenient navigating
through the level because the door acts like a normal door once it the
crystal unlocks it.

This is implemented by adding a property to hiddendoortrigger called
"saveable" that when set to true, sets the variable
"doortriggers.sprite" to true, indicating that the crystal has been
used.

The saveable property could be used in other levels to make doors have
"permanence" although I have only tested it on
caste-hawkthorne-entrance. This functionality makes less since in
valley-chili-fields-2 however because the door is truly "hidden".