Updated saving to all hiddendoortriggers
Now there is no option for "saveable" door triggers as in Niamu's
suggestions. This means that all doors that are opened or shown with
hidden door triggers should now include a property called "trigger"
which indicates the trigger that it is used by. This is needed to
make doors open as soon as the level is rendered.

Affected triggers include:
* Dean's Office 2 (dalmatian)
* House (painting)
* Chili Fields (skull)
* Castle Hawkthorne Entrance (pedestal)

New hidden door triggers should now have that property or else the
loading from save feature won't load properly (the door will need to be
"fixed again").