test-maps: Add reachability test
The first part checks all levels in the *maps* directory and prints warning for non-accesible levels.
There are currently 3 levels that are not reachable via doors: *mountain_pass*, *oriental-mountains*, *oriental-temple* are work in progress.
*new-abedtown* is reachable only from overworld, add it to the default queue, so displayed warnings are correct.

The second part checks few hand-picked levels and fails the test if the level is not reachable.
(mostly towns and levels with cutscenes): *studyroom*, *forest*, *town*, *valley-tacotown*, *valley-3*, *rainbow-bar*, *seabluff*, *castle-hawkthorne-throne*, *greendale-exterior*.