Burn Buildings Fix
We don't have platforms to ghost or search for points within platform shapes anymore. This fixes the collision tiles to be removed when a building is burned.

It also "fixes" the crash when we attempt to find a point inside the old platform shape for the fire to spawn in. There are no platform shapes anymore, so we just add fire blindly and it will appear in wrong areas above the roof. We'll need to develop a proper fix for this later.