Replay - record keystrokes, replay them later
Recording is started/stopped by *F7*, replay by *F8*; the transcript is saved into/read from *replay.json* in the saves folder.
There is an indicator (blinking *R*) in the top left corner when recording/replay is active.
New command line options for it - *--record* and *--replay* that start the recording/replay right away.
Recording is stopped in the case of an error.
Debug tools keys (*F5*,*F6*,*F7*,*F8*) are not recorded and can be used even during replay.